Assignment complete

+++ AssignmentApi/Data/Record.cs
@@ -3,9 +3,9 @@
     public class Record
     {
         public int Id { get; set; }
-        public decimal Velocity { get; set; }
-        public decimal Mass { get; set; }
-        public decimal Energy { get; set; }
+        public double Velocity { get; set; }
+        public double Mass { get; set; }
+        public double Energy { get; set; }
         public string ImpactResult { get; set; }
     }
 }
